+++ src/tracing/ipc/service/consumer_ipc_service.cc
@@ -102,8 +102,8 @@ void ConsumerIPCService::Flush(const protos::FlushRequest& req,
     if (weak_this)
       weak_this->OnFlushCallback(success, std::move(it));
   };
-  GetConsumerForCurrentRequest()->service_endpoint->Flush(
-      static_cast<int>(req.timeout_ms()), std::move(callback));
+  GetConsumerForCurrentRequest()->service_endpoint->Flush(req.timeout_ms(),
+                                                          std::move(callback));
 }
 
 // Called by the service in response to a service_endpoint->Flush() request.
